chore(height,width): 🤖 move to dimensions directory
