+++ src/index.js
@@ -36,7 +36,7 @@ export { default as gridRowStart } from './lib/grid/gridRowStart'
 export { default as gridTemplate } from './lib/grid/gridTemplate'
 export { default as gridTemplateColumns } from './lib/grid/gridTemplateColumns'
 export { default as gridTemplateRows } from './lib/grid/gridTemplateRows'
-export { default as height } from './lib/height'
+export { default as height } from './lib/dimensions/height'
 export { default as justifyContent } from './lib/alignment/justifyContent'
 export { default as justifySelf } from './lib/alignment/justifySelf'
 export { default as margin } from './lib/margin/margin'
@@ -49,4 +49,4 @@ export { default as paddingTop } from './lib/padding/paddingTop'
 export { default as paddingLeft } from './lib/padding/paddingLeft'
 export { default as paddingBottom } from './lib/padding/paddingBottom'
 export { default as paddingRight } from './lib/padding/paddingRight'
-export { default as width } from './lib/width'
+export { default as width } from './lib/dimensions/width'
